fix: modules and permissions
